+++ NEWS.md
@@ -32,6 +32,8 @@
   - `when_any()` (#306)
   - `when_all()` (#306)
 
+* `separate()` now supports regex in the `sep` argument (#320).
+
 
 ## Other changes